feat(toast): ajout d'un hook useToast pour gérer les notifications et suppression du composant Toast
